skill(research): update investigration workflow
