+++ .gemini/commands/investigate.toml
