+++ TODO.md
@@ -165,6 +165,8 @@ notes:
 ## Tempting integrations
 - [ ] Implement error logging and monitoring (e.g., Sentry integration).
 - [ ] Integrate with Rema (Æ), Coop Mega (Member) and Trumf
+    - [ ] Support manual uploading of receipts (e.g., via image or CSV) to extract purchase data and automatically update inventory.
+    - [Helpful openai overview](https://gist.github.com/HelgeSverre/80a7f34f874336324184a0c513c2e6a2)
 
 ## Commercialization & Next Steps
 - [ ] Choose and implement a "Source Available" license (e.g., PolyForm Noncommercial or BSL 1.1) to prevent unauthorized financial benefit by third parties.
