+++ archive/agents/researcher.agent.md
