+++ research/api_legality/overview.md
@@ -1,0 +1,11 @@
+# Research Overview: raw_data
+
+This file provides a rational structure and links to processed research data. Redundant information has been filtered out.
+
+### info
+- **Source:** [info.md](info.md)
+- **Summary:** # Legal Context of Reverse Engineering and Private APIs
+
+## Core Legal Frameworks
+
+The legality of reverse engineering private or undocumented APIs is a complex intersection of **Copyright Law**, **Co...
