+++ research/coop-integration/overview.md
@@ -1,0 +1,10 @@
+# Research Overview: coop-integration
+
+This file provides a rational structure and links to processed research data. Redundant information has been filtered out.
+
+### info
+- **Source:** [info.md](info.md)
+- **Summary:** # Coop Norway (Coop Mega) App API Integration
+
+## Background
+Coop Norge (which includes Coop Mega, Obs, Extra, etc.) does not provide a public "Membership" or "Grocery Receipt" API for external develo...
