+++ research/grocery_private_apis/overview.md
@@ -1,0 +1,10 @@
+# Research Overview: grocery_private_apis
+
+This file provides a rational structure and links to processed research data. Redundant information has been filtered out.
+
+### info
+- **Source:** [info.md](info.md)
+- **Summary:** # Raw Data from Gist & Search: Helge Sverre Private APIs
+
+## Gist: 80a7f34f874336324184a0c513c2e6a2
+The gist contains reverse-engineered API documentation for the three major Norwegian grocery chains:...
