+++ research/rema-ae-integration/overview.md
@@ -1,0 +1,10 @@
+# Research Overview: rema-ae-integration
+
+This file provides a rational structure and links to processed research data. Redundant information has been filtered out.
+
+### info
+- **Source:** [info.md](info.md)
+- **Summary:** # Rema 1000 Æ API Integration
+
+## Background
+The Rema 1000 "Æ" app uses an internal backend API. This API is not officially public, but has been extensively reverse-engineered by the community, notabl...
