+++ research/whatibuy-integration/overview.md
@@ -1,0 +1,10 @@
+# Research Overview: whatibuy-integration
+
+This file provides a rational structure and links to processed research data. Redundant information has been filtered out.
+
+### info
+- **Source:** [info.md](info.md)
+- **Summary:** # WhatIBuy - Grocery Integration Research
+
+## Background
+WhatIBuy (operated by the Danish company Kauza ApS / What I Buy ApS) is a digital service and mobile app available in Denmark (`whatibuy.dk`), ...
